Fix OffersController and OffersDao not inserting offers

+++ HTMLPure/src/main/java/it/polimi/webapp/beans/OffersPageArgs.java
@@ -3,7 +3,7 @@ package it.polimi.webapp.beans;
 import it.polimi.webapp.dao.OffersDao;
 import org.jetbrains.annotations.Nullable;
 
-public record OffersPageArgs(@Nullable Integer auctionId, @Nullable Integer offerPrice, OffersDao.InsertionResult offerInsertion) {
+public record OffersPageArgs(@Nullable Integer auctionId, @Nullable Double offerPrice, OffersDao.InsertionResult offerInsertion) {
 
     public OffersPageArgs() {
         this(null, null, OffersDao.InsertionResult.DONE);
